Split Modify Class. Create the name space Modify with Curve and Surface. Implemented the method into NurbsBase and KnotVector
