+++ src/GShark.Test.XUnit/Modify/SurfaceTests.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace GShark.Test.XUnit.Modify
+{
+    public class SurfaceTests
+    {
+    }
+}
