+++ src/GShark.Test.XUnit/Sampling/CurveTests.cs
@@ -116,7 +116,7 @@ namespace GShark.Test.XUnit.Sampling
         public void Divide_By_Number_Of_Segments_Returns_Points_And_Parameters_Along_Curve()
         {
             // Arrange
-            NurbsBase curve = NurbsBaseCollection.NurbsBasePlanarExample();
+            NurbsBase curve = NurbsBaseCollection.NurbsPlanarExample();
             double[] tValuesExpected = {
                     0,
                     0.12294074023135007,
@@ -147,7 +147,7 @@ namespace GShark.Test.XUnit.Sampling
         public void Divide_By_Length_Returns_Points_And_Parameters_Along_Curve()
         {
             // Arrange
-            NurbsBase curve = NurbsBaseCollection.NurbsBasePlanarExample();
+            NurbsBase curve = NurbsBaseCollection.NurbsPlanarExample();
             double[] tValuesExpected = {
                 0,
                 0.12294074023135007,
@@ -328,7 +328,7 @@ namespace GShark.Test.XUnit.Sampling
         public void AdaptiveSample_Returns_The_ControlPoints_If_Curve_Has_Grade_One()
         {
             // Arrange
-            List<Point3> locationPts = NurbsBaseCollection.NurbsBasePlanarExample().ControlPointLocations;
+            List<Point3> locationPts = NurbsBaseCollection.NurbsPlanarExample().ControlPointLocations;
             NurbsCurve curve = new NurbsCurve(locationPts, 1);
 
             // Act
@@ -383,7 +383,7 @@ namespace GShark.Test.XUnit.Sampling
         public void AdaptiveSample_Use_MaxTolerance_If_Tolerance_Is_Set_Less_Or_Equal_To_Zero()
         {
             // Act
-            (List<double> tValues, List<Point3> pts) = Curve.AdaptiveSample(NurbsBaseCollection.NurbsBasePlanarExample());
+            (List<double> tValues, List<Point3> pts) = Curve.AdaptiveSample(NurbsBaseCollection.NurbsPlanarExample());
 
             // Assert
             tValues.Should().NotBeEmpty();
